requirements.txt
# Core Streamlit Framework
streamlit>=1.46.0

# Data Processing
pandas>=2.3.0
numpy>=2.3.1

# Snowflake Integration
snowflake-snowpark-python>=1.33.0
snowflake-connector-python>=3.15.0

# Data Visualization (if needed for future enhancements)
altair>=5.5.0

# Additional Utilities
pyyaml>=6.0.2
python-dateutil>=2.9.0

# Development and Testing (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0 